32 bit test was only being done on python 3.5

+++ .ci/run_32bit_test.sh
@@ -4,19 +4,32 @@
 # to be called in any other manner.
 #
 
+PYV=$PYTHON_VERSION
+
 apt-get update
 apt-get install -y libssl-dev openssl wget build-essential
 cd /
-wget https://www.python.org/ftp/python/3.5.3/Python-3.5.3.tgz
-tar xf Python-3.5.3.tgz
-cd Python-3.5.3
-./configure --prefix=/python35
+wget https://www.python.org/ftp/python/$PYV/Python-$PYV.tar.xz
+tar xf Python-$PYV.tar.xz
+cd Python-$PYV
+./configure --prefix=/mypython
 make && make install
-cd /python35/bin
-ln -s python3.5 python
-ln -s pip3 pip
+cd /mypython/bin
 
-export PATH=/python35/bin:$PATH
+if [ ! -e python ]; then
+    ln -s python3 python
+fi
+
+if [ ! -e pip ]; then
+    if [ -e pip3 ]; then
+        ln -s pip3 pip
+    else
+        wget https://bootstrap.pypa.io/get-pip.py
+        ./python get-pip.py
+    fi
+fi
+
+export PATH=/mypython/bin:$PATH
 
 pip install numpy pytest cython
